+++ src/Machina/Descriptors/Components/RobotJoint.cs
@@ -17,38 +17,32 @@ namespace Machina.Descriptors.Components
     /// <summary>
     /// Represents a physical Joint between two axes.
     /// </summary>
-
-    /// <summary>
-    /// Defines the type of RobotJoint
-    /// </summary>
-    public enum RobotJointType {
-        /// <summary>
-        /// Joint moves linearly along Z axis.
-        /// </summary>
-        Linear,
-
-        /// <summary>
-        /// Joint rotates around Z axis.
-        /// </summary>
-        Revolute
-    };
-    
     internal class RobotJoint
     {
         /// <summary>
         /// The base plane for this joint's transformation (linear or rotary).
         /// </summary>
-        Matrix4x4 BasePlane { get; set; }
-        
+        internal Matrix4x4 BasePlane { get; set; }
+
         /// <summary>
         /// The end plane this joint transforms. 
         /// </summary>
-        Matrix4x4 TransformedPlane { get; set; }
+        internal Matrix4x4 TransformedPlane { get; set; }
 
         /// <summary>
         /// Is this a linear a revolution joint?
         /// </summary>
-        RobotJointType RobotJointType { get; set; }
+        internal RobotJointType RobotJointType { get; set; }
+
+        /// <summary>
+        /// The linear/angular limitations for this joint.
+        /// </summary>
+        internal Interval JointRange { get; set; }
+
+        /// <summary>
+        /// Max joint speed in degrees/sec or mm/sec.
+        /// </summary>
+        internal double MaxSpeed { get; set; }
         
         // mesh if applicable?
     }