+++ src/Machina/Descriptors/Models/RobotModelBase.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Machina.Descriptors.Models
+{
+    /// <summary>
+    /// A base class for the description of machine models.
+    /// </summary>
+    abstract class RobotModelBase
+    {
+
+    }
+}